refactor: separate messagerender logic from 'Message'
